Fix crash if info dictionary is NSNull
If the info dictionary passed to the initialiser is NSNull then the dictionary lookup for the key kNextURL will cause a crash, as the method for it is undefined on NSNull. Doing the checks using the logical and will only perform the lookup if there actually is a proper dictionary passed.